requirements.txt
customtkinter==5.2.2
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
schedule==1.2.1
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pyinstaller==6.3.0
pillow==10.1.0